+++ react/components/ActionMenu/index.js
@@ -151,6 +151,8 @@ ActionMenu.propTypes = {
       label: PropTypes.node,
       onClick: PropTypes.func,
       disabled: PropTypes.bool,
+      /** Optional testid property */
+      testId: PropTypes.string,
       /** whether option has inline toggle */
       toggle: PropTypes.shape({
         checked: PropTypes.bool,