+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `disabled` prop for ActionMenu items
 - Hide clear button on `InputSearch` when disabled
+- `testId` property to option in `ActionMenu`.
 
 ### Changed